added onTriggerenter for forcefield also checks state of demon boss
added onTriggerenter for forcefield
Forcefield checks for state of demon boss not inactive. ie player is near boss before activation.
Also changed the enamy that is spawned to bat - thought this worked better given we are up in the air.
Hooked up probability of bat and ghost spawners to difficulty level.